style: 디렉토리 구조 MVC 패턴화
